+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### Version 0.2 (2014-10-30)
+
+  - BREAKING CHANGES: plugins!
+  - feat: smart MagicParser
+  - feat: CoParser utility
+
 ### Version 0.1 (2014-08-05)
 
   - feat add support for SRT file format